+++ contracts/treasury/src/state.rs
@@ -1,12 +1,12 @@
-use cosmwasm_std::{Addr, Uint128};
+use cosmwasm_std::{Addr, Decimal, Uint128};
 use cw_storage_plus::Item;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct Config {
-    /// Distribution rate in percents (0-100) which goes to distribution contract
-    pub distribution_rate: u8,
+    /// Distribution rate (0-1) which goes to distribution contract
+    pub distribution_rate: Decimal,
     /// Address of distribution contract, which will receive funds defined but distribution_rate %
     pub distribution_contract: Addr,
     /// Address of reserve contract, which will receive funds defined by 100-distribution_rate %
